# Full support for acceleration and rotation. Optional on-device copyBits and fullscreen text scrolling.