@makermelissa 2025-05-04
Thanks for your contribution. I referenced your rotation code in #34, but would prefer to leave the display specific optimizations out in favor of library simplicity. Perhaps you could subclass this library and make an accelerated version.